test(server-core): colocate spectator wire guard coverage
